requirements.txt
mysql-connector-python
pygame
pydub
